+++ drivers/disk/sdmmc_stm32.c
@@ -10,7 +10,7 @@
 #include <drivers/disk.h>
 #include <drivers/clock_control.h>
 #include <drivers/clock_control/stm32_clock_control.h>
-#include <pinmux/stm32/pinmux_stm32.h>
+#include <pinmux/pinmux_stm32.h>
 #include <drivers/gpio.h>
 #include <logging/log.h>
 #include <soc.h>
