+++ drivers/flash/flash_stm32_qspi.c
@@ -13,7 +13,7 @@
 #include <arch/common/ffs.h>
 #include <sys/util.h>
 #include <soc.h>
-#include <pinmux/stm32/pinmux_stm32.h>
+#include <pinmux/pinmux_stm32.h>
 #include <drivers/clock_control/stm32_clock_control.h>
 #include <drivers/clock_control.h>
 #include <drivers/flash.h>
