+++ drivers/gpio/gpio_stm32.h
@@ -12,7 +12,7 @@
  */
 
 #include <drivers/clock_control/stm32_clock_control.h>
-#include <pinmux/stm32/pinmux_stm32.h>
+#include <pinmux/pinmux_stm32.h>
 #include <drivers/gpio.h>
 
 /* GPIO buses definitions */
