+++ drivers/i2s/i2s_ll_stm32.c
@@ -15,7 +15,7 @@
 #include <stm32_ll_spi.h>
 #include <drivers/clock_control/stm32_clock_control.h>
 #include <drivers/clock_control.h>
-#include <pinmux/stm32/pinmux_stm32.h>
+#include <pinmux/pinmux_stm32.h>
 
 #include "i2s_ll_stm32.h"
 #include <logging/log.h>
