+++ drivers/pinmux/CMakeLists.txt
@@ -12,5 +12,5 @@ zephyr_sources_ifdef(CONFIG_PINMUX_MCUX_LPC        pinmux_mcux_lpc.c)
 zephyr_sources_ifdef(CONFIG_PINMUX_RV32M1          pinmux_rv32m1.c)
 zephyr_sources_ifdef(CONFIG_PINMUX_SAM0            pinmux_sam0.c)
 zephyr_sources_ifdef(CONFIG_PINMUX_SIFIVE          pinmux_sifive.c)
-zephyr_sources_ifdef(CONFIG_PINMUX_STM32           stm32/pinmux_stm32.c)
+zephyr_sources_ifdef(CONFIG_PINMUX_STM32           pinmux_stm32.c)
 zephyr_sources_ifdef(CONFIG_PINMUX_XEC             pinmux_mchp_xec.c)
