+++ drivers/spi/spi_ll_stm32.c
@@ -21,7 +21,7 @@ LOG_MODULE_REGISTER(spi_ll_stm32);
 #include <dt-bindings/dma/stm32_dma.h>
 #include <drivers/dma.h>
 #endif
-#include <pinmux/stm32/pinmux_stm32.h>
+#include <pinmux/pinmux_stm32.h>
 #include <drivers/clock_control/stm32_clock_control.h>
 #include <drivers/clock_control.h>
